Refactor and enhance volume handling, add OBV support
Refactored `PriceVol` and related classes to replace `Vol` with `Volume` (using `Long` for larger datasets). Removed `PriceVolAsClass` and merged its functionality into `PriceVol`. Introduced the `FilterOBV` class for On-Balance Volume calculations. Updated `RecordPrices` and `StockPriceVol` to handle `Volume` consistently. Improved serialization, error handling, and logging. Removed deprecated code and optimized performance for large datasets.

'Fixed an issue with teh graph RangeValue calculator to retunr a non null data range with a distance of zero. This was causing a crash in the graph engine that you could not recover.